+++ oneview_redfish_toolkit/app.py
@@ -64,6 +64,7 @@ from oneview_redfish_toolkit.blueprints.network_port_collection \
     import network_port_collection
 from oneview_redfish_toolkit.blueprints.odata import odata
 from oneview_redfish_toolkit.blueprints.redfish_base import redfish_base
+from oneview_redfish_toolkit.blueprints.resource_block import resource_block
 from oneview_redfish_toolkit.blueprints.resource_block_collection \
     import resource_block_collection
 from oneview_redfish_toolkit.blueprints.service_root import service_root
@@ -129,6 +130,7 @@ def main(config_file_path, logging_config_file_path):
     app.register_blueprint(network_port)
     app.register_blueprint(session)
     app.register_blueprint(resource_block_collection)
+    app.register_blueprint(resource_block)
     app.register_blueprint(zone_collection)
     app.register_blueprint(zone)